+++ app/src/main/java/za/co/opsmobile/coindispense/dispense/store/DispenseStoreActions.java
@@ -1,16 +1,17 @@
 package za.co.opsmobile.coindispense.dispense.store;
 
+import java.math.BigDecimal;
 import java.util.HashMap;
 
 /**
  * Created by Daniel Oosthuizen on 2015/09/05.
  */
 public interface DispenseStoreActions {
-    void initialise(Float[] validDenominations);
+    void initialise(BigDecimal[] validDenominations);
 
-    void addPayment(Float denomination);
+    void addPayment(BigDecimal denomination);
 
-    void setChange(HashMap<Float, Integer> changeTransaction);
+    void setChange(HashMap<BigDecimal, Integer> changeTransaction);
 
-    void setCost(float cost);
+    void setCost(BigDecimal cost);
 }
